Remove passing identity value areound as a BigDecimal since the underlying value is a long. Consistent naming for DataValueFactory methods related to DECIMAL. JSR169 releted cleanup.
git-svn-id: https://svn.apache.org/repos/asf/incubator/derby/code/trunk@125423 13f79535-47bb-0310-9956-ffa450edef68

+++ java/engine/org/apache/derby/iapi/sql/conn/LanguageConnectionContext.java
@@ -54,7 +54,6 @@ import org.apache.derby.catalog.UUID;
 import java.util.Locale;
 import java.util.Hashtable;
 import java.util.Vector;
-import java.math.BigDecimal;
 
 /**
  * LanguageConnectionContext keeps the result sets,
@@ -515,7 +514,7 @@ public interface LanguageConnectionContext extends Context {
 	 *
 	 * @return the generated identity column value
 	 */
-	public BigDecimal getIdentityValue();
+	public Long getIdentityValue();
 
 	/**
 	 * Set the field of most recently generated identity column value.
